+++ src/BottleOfWine.java
@@ -1,5 +1,5 @@
 public class BottleOfWine extends Product {
-    public BottleOfWine(String name, int price) {
+    public BottleOfWine(String name, Long price) {
         super(name, price);
     }
